Add configuration file
Alter munki-trello to read and use a configuration file, using the
python default ConfigParser module style syntax for the configuration
file. This will read the configuration file from a set of default
locations, or one specified on the command line.

A template for the configuration file is provided, with some comments,
and the README.md has been updated with this new option.

In order to make sure that the command line options override the
configuration file, the defaults are now set in the configuration file
parsing, rather than on the command line options.