Add an option to use a single list for production
Make it possible to use a single list for all the packages in production,
rather than a new list for each day. To do this we have added the
``--prod-list`` option (defaults to 'Production'), and if the ``--suffix``
(or ``--prod-suffix``) option is unset, packages will be added to the
production list.

We have also implemented an option to change the date format from the default
``%d/%m/%y`` to a user supplied format. This has meant a need to change from
SafeConfigParser to RawConfigParser (as safe config parser will try to
inteprelate %-encoded values)