Fix a bug with unknown lists
Ignore cards that are in 'unknown' lists (i.e. not ones that are defined in
config file).

Upto this point the assumption was that the software board would only
have software cards added by us. However, in order to allow other
boards (e.g. a list of things to package) we restrict the software to
only consider cards in lits tha have been configured.