Add feature to move between munk repositiories
Allow multiple munki repositories to be configured, and assign each
catalog list a munki repository.

If a package moves between two catalog lists with different repositories
also migrate the package between the two munki repositories.

This is the first version of this feature, and has not been well tested,
but seems to work.