Add option to schedule auto staging
Add two options which allow finer grained control of autostaging by
only autostaging if within a defined timeperiod. There are two extra
configuration file sections: a global schedule and a per-catalog
schedule. You need to meet both critera to auto stage.

Add some tests for the scheduling period options; not that there is
a `short' set of tests and a full set of tests; the full tests take
some time to run (around 1 minute compared with < second). The full
tests are skipped by default.

As well as documenting the new sections, tweak README.md to fix
a minor formating problem.