Configuration file: quoting of trello items
Add code to allow the trello board configuration (`key`, `token` and `boardid`)
to be quoted in the configuration file. As this is not the expected
situation, update the README file to try to encourage people to use
unquoted values. However, following the Robustness principle/Postel's
Law allow for this to happen.

Hopefully this improves the documentation and fixes issues encountered
in ox-it/munki-staging#21