Better detection of no repository
If you use a configuration file, you need to specify a Munki
repository in the configuration file.

In this update we try to make this clearer in the documentation, as it
may have been hidden away before and add a check for this condition to
the main script to try to catch when this happens.

We also address two bugs that occur when no configuration are used.
Firstly, the detection of the fact there was no configuration file is
broken as  RawConfigParser.read returns an empty list and apparently
python thinks that '[] >= 1' is True; we now correctly use the length
of the list for this check. Secondly, the default value for the
repository with no configuration file was being returned as a
dictionary rather than a list, which we have now addressed.