Try to track down bug with trello catalogs not having munki repositories
As reported in ox-it/munki-staging#21 it seems there is a certian
configuration that will not assign a Munki repository to a package or
(more likely) a trello catalog. This commit adds code to try to track
down the source of this error which currently I am not able to
reproduce.