Add a per-pacage override to staging days
Currently the number of days to autostage a package is set in the
configuration file for munki-staging on a per-trello list/munki
catalog basis.

This patch adds the option to also set a per-package number of staging
days by creating a 'munki_staging' key in a pkgsinfo file. It also
adds a python script that will do this. This will override the default
setting.

Note: autostaging must be on for this to have any effect